move put scoring classes. Start reorganizing CarrierScoring - needs to get filled with right content
